refactor(resource section): move root component to pages folder
