+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-20T17:18:34.958Z\n"
-"PO-Revision-Date: 2019-02-20T17:18:34.958Z\n"
+"POT-Creation-Date: 2019-02-21T07:48:10.435Z\n"
+"PO-Revision-Date: 2019-02-21T07:48:10.435Z\n"
 
 msgid "Cancel"
 msgstr ""
